ai_service/requirements.txt
torch
transformers
pillow
flask
torchvision